+++ .idea/compiler.xml
@@ -6,11 +6,11 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="java-lessons" />
+        <module name="bicycles-service" />
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel>
-      <module name="java-lessons" target="1.8" />
+      <module name="bicycles-service" target="1.8" />
     </bytecodeTargetLevel>
   </component>
 </project>
