+++ pom.xml
@@ -4,8 +4,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>ldrzewiecki</groupId>
-    <artifactId>java-lessons</artifactId>
+    <groupId>p76.bicycles</groupId>
+    <artifactId>bicycles-service</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>jar</packaging>
