+++ src/main/java/p76/bicycles/dto/Bicycle.java
@@ -1,8 +1,6 @@
-package ldrzewiecki.dto;
+package p76.bicycles.dto;
 
 import java.util.Arrays;
-import java.util.HashMap;
-import java.util.Map;
 
 public class Bicycle {
